requirements.txt
alembic==0.9.9
aniso8601==3.0.0
atomicwrites==1.1.5
attrs==18.1.0
click==6.7
coverage==4.5.1
flake8==3.5.0
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.2.0
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.3
pycodestyle==2.3.1
pydocstyle==2.1.1
pyflakes==1.6.0
pytest==3.6.0
pytest-cov==2.5.1
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.4
six==1.11.0
snowballstemmer==1.2.1
SQLAlchemy==1.2.7
Werkzeug==0.14.1
